+++ miden-tx/src/tests/kernel_tests/test_note.rs
@@ -403,13 +403,10 @@ fn test_get_note_serial_number() {
     // calling get_serial_number should return the serial number of the note
     let code = "
         use.kernel::prologue
-        use.kernel::note->note_internal
-        use.kernel::note
+        use.miden::note
 
         begin
             exec.prologue::prepare_transaction
-            exec.note_internal::prepare_note
-            dropw dropw dropw dropw
             exec.note::get_serial_number
 
             # truncate the stack
@@ -519,4 +516,30 @@ fn test_get_inputs_hash() {
     expected_stack.extend_from_slice(&expected_5);
 
     assert_eq!(process.get_stack_state()[0..16], expected_stack);
+}
+
+#[test]
+fn test_get_current_script_hash() {
+    let tx_context = TransactionContextBuilder::with_standard_account(ONE)
+        .with_mock_notes_preserved()
+        .build();
+
+    // calling get_script_hash should return script hash
+    let code = "
+    use.kernel::prologue
+    use.miden::note
+
+    begin
+        exec.prologue::prepare_transaction
+        exec.note::get_script_hash
+
+        # truncate the stack
+        swapw dropw
+    end
+    ";
+
+    let process = tx_context.execute_code(code).unwrap();
+
+    let script_hash = tx_context.input_notes().get_note(0).note().script().hash();
+    assert_eq!(process.stack.get_word(0), script_hash.as_elements());
 }