+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Implemented new `digest_from_hex!` macro (#984).
 - Added Format Guidebook to the `miden-lib` crate (#987).
 - Added conversion from `Account` to `AccountDelta` for initial account state representation as delta (#983).
+- [BREAKING] Added `miden::note::get_script_hash` procedure (#995).
 
 ## 0.6.2 (2024-11-20)